+++ test/rmgpy/solver/liquidTest.py
@@ -148,7 +148,7 @@ class LiquidReactorCheck:
 
         rxn_system.initialize_model(core_species, core_reactions, edge_species, edge_reactions)
 
-        tlist = np.array([10 ** (i / 10.0) for i in range(-130, -49)], np.float64)
+        tlist = np.array([10 ** (i / 10.0) for i in range(-130, -49)], float)
 
         # Integrate to get the solution at each time point
         t, y, reaction_rates, species_rates = [], [], [], []
@@ -162,9 +162,9 @@ class LiquidReactorCheck:
             species_rates.append(rxn_system.core_species_rates.copy())
 
         # Convert the solution vectors to np arrays
-        t = np.array(t, np.float64)
-        reaction_rates = np.array(reaction_rates, np.float64)
-        species_rates = np.array(species_rates, np.float64)
+        t = np.array(t, float)
+        reaction_rates = np.array(reaction_rates, float)
+        species_rates = np.array(species_rates, float)
 
         # Check that we're computing the species fluxes correctly
         for i in range(t.shape[0]):
@@ -599,7 +599,7 @@ class LiquidReactorCheck:
 
         rxn_system.initialize_model(core_species, core_reactions, edge_species, edge_reactions)
 
-        tlist = np.array([10 ** (i / 10.0) for i in range(-130, -49)], np.float64)
+        tlist = np.array([10 ** (i / 10.0) for i in range(-130, -49)], float)
 
         # Integrate to get the solution at each time point
         t, y, reaction_rates, species_rates = [], [], [], []
